Authors HTML generation issue
I have found out that when you do not define the authorURL, or affiliationURL inside the d-front-matter tag, it tries to create a DIV element within a P tag, which is not allowed by the HTML standard. As a result, the display of the authors and affiliations is incorrect. See https://github.com/distillpub/template/issues/80 for a full description.